Put very general share redemption code into common
