+++ common/new-bet.ts
@@ -20,9 +20,9 @@ import { noFees } from './fees'
 import { addObjects } from './util/object'
 import { NUMERIC_FIXED_VAR } from './numeric-constants'
 
-export type CandidateBet<T extends Bet> = Omit<T, 'id' | 'userId'>
+export type CandidateBet<T extends Bet = Bet> = Omit<T, 'id' | 'userId'>
 export type BetInfo = {
-  newBet: CandidateBet<Bet>
+  newBet: CandidateBet
   newPool?: { [outcome: string]: number }
   newTotalShares?: { [outcome: string]: number }
   newTotalBets?: { [outcome: string]: number }
@@ -46,7 +46,7 @@ export const getNewBinaryCpmmBetInfo = (
   const probBefore = getCpmmProbability(pool, p)
   const probAfter = getCpmmProbability(newPool, newP)
 
-  const newBet: CandidateBet<Bet> = {
+  const newBet: CandidateBet = {
     contractId: contract.id,
     amount,
     shares,
@@ -96,7 +96,7 @@ export const getNewBinaryDpmBetInfo = (
   const probBefore = getDpmProbability(contract.totalShares)
   const probAfter = getDpmProbability(newTotalShares)
 
-  const newBet: CandidateBet<Bet> = {
+  const newBet: CandidateBet = {
     contractId: contract.id,
     amount,
     loanAmount,
@@ -133,7 +133,7 @@ export const getNewMultiBetInfo = (
   const probBefore = getDpmOutcomeProbability(totalShares, outcome)
   const probAfter = getDpmOutcomeProbability(newTotalShares, outcome)
 
-  const newBet: CandidateBet<Bet> = {
+  const newBet: CandidateBet = {
     contractId: contract.id,
     amount,
     loanAmount,
